backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Marshmallow
marshmallow-sqlalchemy
Flask-CORS
Flask-Mail
Celery
redis
boto3
Pillow
psycopg2-binary
flasgger
pytest
python-dotenv
gunicorn
flask-socketio
eventlet
